[michocoq] Avoid redefining `False` in `util.v`
The name of the Michelson data constructor `False` clashes with the
Coq proposition for contradiction. To avoid qualifying the Coq
proposition, `util.v` redefines it so `util.False` is defined as
`Logic.False`. This has unpleasant side effects (see
https://gitlab.com/nomadic-labs/mi-cho-coq/-/issues/53) so we remove
this definition and qualify the name in the few places where it is
used instead.

Closes #53.