[Mi-Cho-Coq] Distinguish entrypoint trees from types
This is a first step toward removing annotations in Michelson
types. The only place where these annotations are semantically
relevant is when they are used to build entrypoint trees whose nodes
are `or` types and leaves are all other types. This commit defines a
new type `entrypoints.entrypoint_tree` to track the types that are
used as entrypoint trees (those that represent contract parameters).